requirements.txt
Django>=2.2.9,<3.0.0
djangorestframework>=3.11.0,<3.12.0